added clientserver.go
The idea is:

You launch the agent on a host with a "manifest" of command to run.
The agent is :
- parsing all the commands from the manifest and add the to a stack
- creating a communication socket
- starts to process the stack in FIFO mode
- when there is a request on the socket, it reads it and resturns its
  current working status

If the agent is launched a second time (or many other time), it checks
for the presence of the socket file and the ability to dial in.

If it can dial in, he pushes the content of its manifest to the stack of
the listening process.