Corrected a bug in the relink function. The end node of a subtask was linked to the start node of the same subtask.
The problem was that I was setting 1 to the couple (sourceOrigin, DestInitialSubtask) and when I was adding the two rows for the subtask exit nodes, I was adding the 1 value previously set.
not clear ? sorry :)

The implementation is storing the couples of coordinates in a slice and comparing it.
Not efficient at all... May be fixe someday